+++ src/main/java/hibi/boathud/MenuInteg.java
@@ -85,6 +85,13 @@ public class MenuInteg implements ModMenuApi {
 				.setSaveConsumer(newVal -> Config.minimapLockNorth = newVal)
 				.build())
 
+		.addEntry(entryBuilder.startEnumSelector(MINIMAP_SHAPE, MinimapShape.class, MinimapShape.values()[Config.minimapShape])
+				.setDefaultValue(MinimapShape.CIRCLE)
+				.setTooltip(TIP_MINIMAP_SHAPE)
+				.setSaveConsumer(newVal -> Config.minimapShape = newVal.ordinal())
+				.setEnumNameProvider(value -> Text.translatable("boathud.option.minimap_shape." + value.toString()))
+				.build())
+
 			.addEntry(entryBuilder.startIntField(MINIMAP_ICE_DETECTION_RANGE, Config.minimapIceDetectionRange)
 				.setDefaultValue(15)
 				.setMin(1)
@@ -176,6 +183,9 @@ public class MenuInteg implements ModMenuApi {
 	public enum BarType {
 		PACKED, MIXED, BLUE
 	}
+	public enum MinimapShape {
+		CIRCLE, SQUARE
+	}
 		public enum SpeedFormat {
 		MS, KMPH, MPH, KT
 	}
@@ -197,6 +207,8 @@ public class MenuInteg implements ModMenuApi {
 		TIP_MINIMAP_SHOW_ALL_HEIGHTS = Text.translatable("boathud.tooltip.minimap_show_all_heights"),
 		MINIMAP_LOCK_NORTH = Text.translatable("boathud.option.minimap_lock_north"),
 		TIP_MINIMAP_LOCK_NORTH = Text.translatable("boathud.tooltip.minimap_lock_north"),
+		MINIMAP_SHAPE = Text.translatable("boathud.option.minimap_shape"),
+		TIP_MINIMAP_SHAPE = Text.translatable("boathud.tooltip.minimap_shape"),
 	
 	MINIMAP_ICE_DETECTION_RANGE = Text.translatable("boathud.option.minimap_ice_detection_range"),
 	TIP_MINIMAP_ICE_DETECTION_RANGE = Text.translatable("boathud.tooltip.minimap_ice_detection_range"),